requirements.txt
groq
requests
folium
python-dotenv
fastapi
uvicorn[standard]
python-multipart
